+++ wezterm-gui/src/frontend.rs
@@ -144,6 +144,10 @@ impl GuiFrontEnd {
                 false
             }
         });
+        // Re-evaluate the config so that folks that are using
+        // `wezterm.gui.get_appearance()` can have that take effect
+        // before any windows are created
+        config::reload();
         Ok(front_end)
     }